update skeleton form for edit delete and display image

+++ templates/inc/view_index.php
@@ -1,8 +1,8 @@
 <?php
 if(isset($config['url']['action']) && $config['url']['action'] == 'search') {
-    ?>
-    <p><a href="<?=$config['url']['module']?>">Go back</a></p>
-    <?php
+    ?><p><a href="/<?=$config['url']['module']?>">Go back</a></p><?php
+} else {
+    ?><p><a href="/<?=$config['url']['module']?>/add">Add new</a></p><?php
 }
 if(sizeof($data) > 0) {
 
@@ -31,7 +31,7 @@ if(sizeof($data) > 0) {
                     echo "<td class='center'>".$value."</td>";
                 }
                 echo "<td class='center'><a href='".$config['url']['module']."/edit/".current($array_data)."'>Edit</a></td>";
-                echo "<td class='center'><a href='".$config['url']['module']."/delete/".current($array_data)."'>Delete</a></td>";
+                echo "<td class='center'><a href='".$config['url']['module']."/delete/".current($array_data)."' onclick='return confirm(\"Are you sure want to delete this data?\");'>Delete</a></td>";
             echo "</tr>";
         }
